Refactor content repository docker actions.
