+++ src/utils/DockerUtil.js
@@ -6,8 +6,8 @@ import path from 'path';
 import dockerode from 'dockerode';
 import _ from 'underscore';
 import util from './Util';
-import contentRepositoryActions from '../actions/ContentRepositoryActions';
 import Promise from 'bluebird';
+import ContentRepositoryActions from '../actions/ContentRepositoryActions';
 
 export default {
   host: null,
@@ -169,7 +169,7 @@ export default {
                 return;
               }
 
-              contentRepositoryActions.containerCompleted({container});
+              ContentRepositoryActions.containerCompleted({container});
             });
           }
         }
@@ -307,85 +307,6 @@ export default {
         }
       });
     });
-  },
-
-  launchServicePod (repo) {
-    let contentParams = {
-      Env: [
-        "NODE_ENV=development",
-        "STORAGE=memory",
-        "ADMIN_APIKEY=supersecret",
-        "CONTENT_LOG_LEVEL=debug",
-        "CONTENT_LOG_COLOR=true"
-      ],
-      HostConfig: {
-        PublishAllPorts: true,
-        ReadonlyRootfs: true
-      }
-    };
-
-    let presenterParams = {
-      Volumes: {
-        "/var/control-repo": {}
-      },
-      Env: [
-        "NODE_ENV=development",
-        "CONTROL_REPO_PATH=/var/control-repo",
-        "CONTENT_SERVICE_URL=http://content:8080/",
-        "PRESENTER_LOG_LEVEL=debug",
-        "PRESENTER_LOG_COLOR=true",
-        "PRESENTED_URL_DOMAIN=developer.rackspace.com",
-      ],
-      HostConfig: {
-        Binds: [ repo.controlRepositoryLocation + ":/var/control-repo:ro" ],
-        Links: [ "content-" + repo.id + ":content" ],
-        PublishAllPorts: true,
-        ReadonlyRootfs: true
-      }
-    };
-
-    async.series([
-      (cb) => {
-        this.run("content-" + repo.id, "quay.io/deconst/content-service", "latest", contentParams, cb);
-      },
-      (cb) => {
-        this.run("presenter-" + repo.id, "quay.io/deconst/presenter", "latest", presenterParams, cb);
-      }
-    ], (error, containers) => {
-      if (error) {
-        contentRepositoryActions.error({repo, error});
-        return;
-      }
-
-      let [contentContainer, presenterContainer] = containers;
-
-      contentRepositoryActions.podLaunched({repo, contentContainer, presenterContainer});
-    });
-  },
-
-  launchPreparer (repo) {
-    let params = {
-      Volumes: {
-        "/usr/control-repo": {}
-      },
-      Env: [
-        "CONTENT_STORE_URL=" + repo.contentURL(),
-        "CONTENT_STORE_APIKEY=supersecret",
-        "TRAVIS_PULL_REQUEST=false"
-      ],
-      HostConfig: {
-        Binds: [repo.contentRepositoryPath + ":/usr/control-repo"],
-        ReadonlyRootfs: true
-      }
-    };
-
-    this.run("preparer-" + repo.id, "quay.io/deconst/preparer-" + repo.preparer, "latest", params, (error, container) => {
-      if (error) {
-        contentRepositoryActions.error({repo, error});
-        return;
-      }
-
-      contentRepositoryActions.preparerLaunched({repo, container});
-    })
   }
+
 };
